+++ CHANGELOG.md
@@ -4,6 +4,12 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 
 ## Unreleased
 
+* App Release Process:
+   * Added privacy policies to validate and publish
+
+* Application Checklist
+   BugFix - Checklist layout issue fix
+
 
 ## 1.3.0 RC1